+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - popup: Dashboard was not active on non-bandcamp pages
+- popup: Console was broken
 
 ## 0.11.0 (2024-07-29)